+++ src/Ydb.Sdk/CHANGELOG.md
@@ -1,4 +1,10 @@
-- Added 'x-ydb-client-pid' header to any RPC calls
+- Fixed bug `Reader`: unhandled exception in `TryReadRequestBytes(long bytes)`.
+- Handle `YdbException` on `DeleteSession`.
+- Do not invoke `DeleteSession` if the session is not active.
+- `YdbException`: Added cancellation token propagation support in `CommitAsync` and `RollbackAsync`.
+- Deleted legacy exceptions: Driver.TransportException, StatusUnsuccessfulException and InitializationFailureException.
+- Fixed bug: Unhandled exception System.Net.Http.HttpIOException has now been converted to YdbException ([grpc-dotnet issue](https://github.com/grpc/grpc-dotnet/issues/2638)).
+- Added 'x-ydb-client-pid' header to any RPC calls.
 - Added DisableServerBalancer option to ADO.NET session creation; default false.
 
 ## v0.20.1